qemu: assign vfio devices to PCIe addresses when appropriate
Although nearly all host devices that are assigned to guests using
vfio ("<hostdev>" devices in libvirt) are physically PCI Express
devices, until now libvirt's PCI address assignment has always
assigned them addresses on legacy PCI controllers.

This patch tries to assign them to an address on a PCIe controller
instead. First we do some preliminary checks that might allow setting
the flags without doing any extra work, and if those conditions aren't
met (and if libvirt is running privileged so that it has proper
permissions), we perform the (relatively) time consuming task of
reading the device's PCI config to see if it is an Express device. If
this is successful, the connect flags are set based on the result, but
if we aren't able to read the PCI config (most likely due to the
device not being present on the system at the time of the check) we
assume it is (or will be) and Express device, since that is almost
always the case anyway.
Message-Id: <1478548225-1538-18-git-send-email-laine@laine.org>