libxl: add tunnelled migration support
Tunnelled migration doesn't require any extra network connections beside the
libvirt daemon.
It's capable of strong encryption and the default option of openstack-nova.

This patch adds the tunnelled migration(Tunnel3params) support to libxl.
On the src side, the data flow is:
 * libxlDoMigrateSend() -> pipe
 * libxlTunnel3MigrationFunc() polls pipe out and then write to dest stream.

While on the dest side:
Stream -> pipe -> 'recvfd of libxlDomainStartRestore'

The usage is the same as p2p migration, execpt adding one extra '--tunnelled' to
the libvirt p2p migration command.

Signed-off-by: Bob Liu <bob.liu@oracle.com>
Message-Id: <1478837659-3036-1-git-send-email-bob.liu@oracle.com>