logical: Allow [no]overwrite option for poolBuild
https://bugzilla.redhat.com/show_bug.cgi?id=1373711

When using pool-create --build or pool-define and pool-build to create/define
and build/initialize the logical pool, the 'pvcreate' command could fail if
some previous attempt had been used on the source path previously.

So add support for the --override option in order to force usage of the
magic override option "-ff" and the option "-y" in order to force creation
and answer any questions with yes.

NB: Although the reality is part of the process of building the logical
pool involves wiping out the first 512 bytes of the disk block to be
added (e.g. the partition table) because pvcreate requires that. So, to
a degree the device being added is already altered. I suspect the knowlege
of whether the disk was in a vg already could be contained outside the
range of the first 512 bytes.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1479246910-26127-1-git-send-email-jferlan@redhat.com>