Rework existing code for scsi hostdev scsi_host protocol
In review feedback for my vhost-scsi series[1], we encountered a conflict
where existing structs/functions are named "...SCSIHost..." to represent
a hostdev type='scsi' protocol='scsi_host' element.  This is somewhat of
an abbreviation, but jives well with the other elements for this hostdev.
Specifically, a hostdev type='scsi' protocol='iscsi' uses structs and
functions named "...SCSIiSCSI..."

As my series is intending to add a hostdev type='scsi_host', the phrase
"SCSIHost" had to be avoided to prevent conflicts with existing code,
which means the abbreviation gets propogated in new and unplesant ways.

Let's remove this abbreviation and change the existing code from "SCSIHost"
to "SCSISCSIHost" so that a new hostdev type can be added without conflict.
The intention being "<type><protocol>" being used going forward.

[1] End of https://www.redhat.com/archives/libvir-list/2016-November/msg00738.html

Signed-off-by: Eric Farman <farman@linux.vnet.ibm.com>
Message-Id: <1479306360-6899-1-git-send-email-farman@linux.vnet.ibm.com>