logical: Check for existing vgname before building
https://bugzilla.redhat.com/show_bug.cgi?id=1373711

Add a check for an existing volume group name before trying to build
the volume group. Since the process of building a vg involves wiping
the first 512 bytes and using pvcreate on each source device path before
creating the vg - we could conceivably overwrite something that we
shouldn't be. Also, once a pv is part of a vg, the pvcreate would fail
unless we chose to overwrite the volume.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1479306535-8199-1-git-send-email-jferlan@redhat.com>