tests: adapt to gnutls change in dname en-/decoding
A recent change in gnutls that was released with 3.5.6 changed the
behavior of dname en- and decoding to follow RFC4514.

That breaks the related tests which failed validation in
virNetTLSContextCheckCertDNWhitelist due to the strings no more matching
in the fnmatch check.

The fix is a gnutls version dependent definition of the wildcard strings
used by the tests (older gnutls versions require the old order).

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>
Message-Id: <1479309821-9486-2-git-send-email-christian.ehrhardt@canonical.com>