nodedev: Fix crash in libvirtd on vHBA creation path
Providing XML such as:

<device>
  <name>vhba</name>
  <capability type='scsi_host'>
    <capability type='fc_host'>
    </capability>
  </capability>
</device>

would crash libvirt because the '<parent>' isn't a required field, but
for vHBA creation it's expected (day 1 issue - see commit id '81d0ffbc').
The nodedev.rng added in commit id '2c22a68c' has this as an optional field.
NB: On normal udev discovery if a parent field wasn't found, it would be
set to "computer" by udevSetParent, so this is a somewhat unique path.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1479479197-15099-2-git-send-email-jferlan@redhat.com>