util: new function virFileLength()
This new function just calls stat() and returns st_size (or -1 if
there is an error). We may decide we want this function to be more
complex, and handle things like block devices - this is a placeholder
(that works) for any more complicated funtion.

NB: virFileLength() takes a path rather than an fd because it needs to
be called for files that can't be opened (due to permissions).
Message-Id: <20161121050152.5416-2-laine@laine.org>