qemu: Create hugepage path on per domain basis
If you've ever tried running a huge page backed guest under
different user than root, you probably failed. Problem is even
though we have corresponding APIs in the security drivers,
there's no implementation and thus we don't relabel the huge page
path. But even if we did, so far all of the domains share the
same path:

   /hugepageMount/libvirt/qemu

Our only option there would be to set 0777 mode on the qemu dir
which is totally unsafe. Therefore, we can create dir on
per-domain basis, i.e.:

   /hugepageMount/libvirt/qemu/domainName

and chown domainName dir to the user that domain is configured to
run under.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <76aec0f704bbbdcfeae187b75f3c85766259364b.1479818444.git.mprivozn@redhat.com>