qemu: Prevent detaching SCSI controller used by hostdev
Consider a guest started with the following XML snippet:

    <controller type='scsi' model='virtio-scsi' index='0'/>
    <hostdev mode='subsystem' type='scsi'>
      <source>
        <adapter name='scsi_host0'/>
        <address bus='0' target='8' unit='1074151456'/>
      </source>
    </hostdev>

If we attach/create a virtio-scsi hostdev device but forget to include
a virtio-scsi controller, one is silently created for us.  (See
qemuDomainFindOrCreateSCSIDiskController for context.)

Detaching the hostdev, followed by the controller, works well and the
guest behaves appropriately.

If we detach the virtio-scsi controller device first, Libvirt silently
detaches any associated hostdevs for us too.  But all is not well,
as the guest is unable to receive new virtio-scsi devices (the attach
commands succeed, but devices never appear within the guest), nor even
be shutdown, after this point.

It appears that something is tied up in the host virtio layer.
While I investigate this, we can see if a controller is being used by
any hostdevs, and prevent the detach if it would lead us down this path.

  $ virsh detach-device guest_one_virtio_scsi scsicontroller.xml
  error: Failed to detach device from scsicontroller.xml
  error: operation failed: device cannot be detached: device is busy

  $ virsh detach-device guest_one_virtio_scsi scsidisk.xml
  Device detached successfully

  $ virsh detach-device guest_one_virtio_scsi scsicontroller.xml
  Device detached successfully

Signed-off-by: Eric Farman <farman@linux.vnet.ibm.com>
Reviewed-by: Bjoern Walk <bwalk@linux.vnet.ibm.com>
Reviewed-by: Boris Fiuczynski <fiuczy@linux.vnet.ibm.com>
Message-Id: <20161128140122.13472-1-farman@linux.vnet.ibm.com>