util: Introduce virFileRewriteStr
There are couple of places where we have a string and want to
save it to a file. Atomically. In all those places we use
virFileRewrite() but also implement the very same callback which
takes the string and write it into temp file. This makes no
sense. Unify the callbacks and move them to one place.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <17b3ea31a607cefa14a9f3ed56ac35f4d2e7556e.1480932896.git.mprivozn@redhat.com>